+++ src/Pages/DecklistPage/DecklistPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Decklist, DEV_STRING_PRE, GET_DECKLIST, GET_PLACING, GET_TOURNAMENT_NAME } from '../../Data';
+import { Decklist, DEV_STRING_PRE, GET_CASED_ARCHETYPE, GET_EXPERT_DECKLIST } from '../../Data';
 
 import { VDecklist } from '../../Views';
 
@@ -8,13 +8,12 @@ interface IProps {}
 
 export class DecklistPage extends React.Component<IProps> {
     public override componentDidMount(): void {
-        document.title = `${DEV_STRING_PRE}Decklist ${this.username} ${this.tournId}`;
+        document.title = `${DEV_STRING_PRE}Decklist ${this.username} ${GET_CASED_ARCHETYPE(this.archetype)}`;
     }
     
-    private tournId: string;
+    private archetype: string;
     private username: string;
     private decklist?: Decklist;
-    private placing?: string;
 
     public constructor(props: IProps) {
         super(props);
@@ -22,15 +21,15 @@ export class DecklistPage extends React.Component<IProps> {
         if(url.length > 0 && url.charAt(url.length-1) !== "/") {
             url += "/";
         }
-        this.tournId = url.slice(url.indexOf("tournament/")+11, url.indexOf("/decklist/"));
-        this.username = decodeURI(url.slice(url.indexOf("decklist/")+9, url.length-1));
-        this.decklist = GET_DECKLIST(this.tournId, this.username);
-        this.placing = GET_PLACING(this.tournId, this.username);
+        const urlData = decodeURI(url.slice(url.indexOf("decklist/")+9, url.length-1)).split("/");
+        this.archetype = urlData[0];
+        this.username = urlData[1];
+        this.decklist = GET_EXPERT_DECKLIST(this.archetype, this.username);
     }
 
     public render(): React.ReactNode {
-        return ((!this.decklist || !this.placing) ? <>decklist not found</> : <>
-            <VDecklist decklist={this.decklist} title={this.username} subtitle={`${GET_TOURNAMENT_NAME(this.tournId)} ${this.decklist.date}: ${this.placing}`} />
+        return ((!this.decklist) ? <div style={{marginLeft: "4px"}}>decklist not found</div> : <>
+            <VDecklist decklist={this.decklist} title={`${this.username}'s ${GET_CASED_ARCHETYPE(this.archetype)}`} subtitle={""}  />
         </>);
     }
 }