+++ src/Pages/DecklistPage/index.ts
@@ -1,1 +1,2 @@
 export * from "./DecklistPage";
+export * from "./TournamentDecklistPage";