+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,6 @@
-[0.2.5] Overnumber Support
-- id=ttsId typically; printId is on the card
-- removed constant art from CardData
+[1.0.0] Update AllArchetypePage to Tier List
+- Added some "expert" decklists
+  - Split DecklistPage into regular/expert and tournament
+- Added some jest tests
+- Added rough tournament tiers
+- Links to decklists